+++ src/create.js
@@ -1,49 +1,47 @@
-// `onCreate(error, params)` allows custom logic at initialization time.
-// The construction `params` are passed, i.e. can be validated, normalized, etc.
-// `onCreate()` is useful to assign error instance-specific properties.
-//  - Therefore, the default value just assign `params`.
-// Properties that are error type-specific (i.e. same for all instances of that
-// type):
-//  - Should use a separate object where the key is the `error.name` and the
-//    value are the properties
-//  - Those should be used either:
-//     - By letting the error catcher|consumer retrieve object[error.name],
-//       which is preferred since it:
-//        - Decouples error consumption from creation
-//        - Prevents those properties from being printed on the console
-//        - Makes it harder to lose those properties as the error gets passed
-//          around, serialized, wrapped in `error.cause` or `error.errors`, etc.
-//        - Is simpler for error throwing logic
-//     - By assigning them in `onCreate()` using
-//       `Object.assign(this, object[error.name])`
-//        - This is simpler for the error catching|consuming logic
-export const defaultOnCreate = function (error, params) {
-  // eslint-disable-next-line fp/no-mutating-assign
-  Object.assign(error, params)
-}
-
-// When passing parameters to `onCreate()`, ignore keys that:
+// Applied to both `params` and `params.props`.
+// Ignore keys that:
 //  - Would override `Object` prototype (`hasOwnProperty`, etc.) or `Error`
 //    prototype (`toString`)
 //  - Are prototype-specific (`__proto__`, `prototype`, `constructor`)
 //  - Are core error properties (`name`, `message`, `stack`, `cause`, `errors`)
 //  - Are inherited
 //  - Are not enumerable
-export const getOnCreateParams = function (error, params) {
-  const onCreateParams = {}
+// Values that throw when retrieved throw right away.
+// Due to `error.cause`, the second argument should always be a plain object
+// We enforce no third argument since this is cleaner.
+export const normalizeParams = function (error, params) {
+  const paramsA = normalizeObject('second argument', error, params)
+  const props = normalizeObject(
+    'second argument\'s "props"',
+    error,
+    paramsA.props,
+  )
+  return { ...paramsA, props }
+}
+
+const normalizeObject = function (prefix, error, params = {}) {
+  if (!isObject(params)) {
+    throw new TypeError(`Error's ${prefix} must be a plain object: ${params}`)
+  }
+
+  const paramsCopy = {}
 
   // eslint-disable-next-line fp/no-loops
   for (const key of Reflect.ownKeys(params)) {
-    setKey({ error, params, key, onCreateParams })
+    setKey({ error, params, key, paramsCopy })
   }
 
-  return onCreateParams
+  return paramsCopy
 }
 
-const setKey = function ({ error, params, key, onCreateParams }) {
+export const isObject = function (value) {
+  return typeof value === 'object' && value !== null
+}
+
+const setKey = function ({ error, params, key, paramsCopy }) {
   if (isEnum.call(params, key) && !shouldIgnoreKey(error, key)) {
     // eslint-disable-next-line fp/no-mutation, no-param-reassign
-    onCreateParams[key] = params[key]
+    paramsCopy[key] = params[key]
   }
 }
 
@@ -54,4 +52,12 @@ const shouldIgnoreKey = function (error, key) {
   return key in error || IGNORED_KEYS.has(key)
 }
 
-const IGNORED_KEYS = new Set(['errors', 'prototype', 'cause'])
+const IGNORED_KEYS = new Set(['prototype', 'errors', 'cause'])
+
+// `onCreate(error, params)` allows custom logic at initialization time.
+// The construction `params` are passed, i.e. can be validated, normalized, etc.
+// The default value assigns `params.props`.
+export const defaultOnCreate = function (error, { props }) {
+  // eslint-disable-next-line fp/no-mutating-assign
+  Object.assign(error, props)
+}